+++ MonkeyDelivery/Src/Control/States/MenuState.cpp
@@ -82,6 +82,6 @@ void MenuState::moveBox(Vector2D<int> i) {
 		currentSelection = 0;
 	}
 	else if (currentSelection > buttonsUI.size() - 1) {
-		currentSelection = (int)buttonsUI.size()- 1;
+		currentSelection = buttonsUI.size()- 1;
 	}
 }
