+++ MonkeyDelivery/Src/Items/Lantern.cpp
@@ -19,7 +19,7 @@ bool Lantern::useObject() {
 			cout << " Hola linternita circular";
 			player_->changeLState(true);
 			player_->changeFLState(false);
-			sdlutils().soundEffects().at("lantern").setVolume((int)(game_->getSoundEfectsVolume() * game_->getGeneralVolume()));
+			sdlutils().soundEffects().at("lantern").setVolume(game_->getSoundEfectsVolume() * game_->getGeneralVolume());
 			sdlutils().soundEffects().at("lantern").play(0, 1);
 
 			//Flashlight::setActive(false);
@@ -28,7 +28,7 @@ bool Lantern::useObject() {
 			cout << "Adios linternita circular";
 			player_->changeLState(false);
 			player_->setOrientation("off");
-			sdlutils().soundEffects().at("lantern").setVolume((int)(game_->getSoundEfectsVolume() * game_->getGeneralVolume()));
+			sdlutils().soundEffects().at("lantern").setVolume(game_->getSoundEfectsVolume() * game_->getGeneralVolume());
 			sdlutils().soundEffects().at("lantern").play(0, 1);
 
 		}
