+++ MonkeyDelivery/Src/View/Transition.cpp
@@ -19,7 +19,7 @@ void Transition::draw()
 	 /*2- Calculate the size of mask texture. Increment size each frame by a fitting amount.
 	  You would then use this value to calculate the destination rectangle for the sprite. */
 		size += increase;
-		SDL_Rect dest = {(int)(game->getWindowWidth() / 2) - (size / 2), (int)(game->getWindowHeight() / 2) - (size / 2), size, size};
+		SDL_Rect dest = {(game->getWindowWidth() / 2) - (size / 2), (game->getWindowHeight() / 2) - (size / 2), size, size};
 
 	  /*3- Before rendering your actual game objects, call SDL_RenderSetClipRect(renderer, &dest);
 	  This assures that nothing is rendered outside of the(square) circle texture area. That way everything outside of the texture should stay black.*/
